requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
fake-useragent==1.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
outcome==1.2.0
Pillow==9.4.0
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-telegram-bot==20.1
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
telegram==0.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
